requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-api-python-client==2.108.0
google-auth==2.23.4
python-dotenv==1.0.0
python-dateutil==2.8.2
pydantic==2.5.0
